+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Added
 - A maximum height indicator to the level editor grid to signal the peak of the level (default: 800)
 
+### Fixed
+- The level editor ruler text is now properly scaling with the zoom
+
 ## [1.1] 2015-03-28
 ### Added
 - This CHANGELOG